# Fix just a few stacked borrow issues, and improve extend performance